+++ src/tessera/exec/tessera_dump.c
@@ -264,7 +264,7 @@ static void dump_index(FILE* fp, int64_t ref, const char* arr)
 
 static void usage(void)
 {
-    printf("Usage: realm-dump <file> [?][<ref>] [<array>]\n");
+    printf("Usage: tessera-dump <file> [?][<ref>] [<array>]\n");
     exit(1);
 }