+++ src/tessera/object-store/impl/apple/external_commit_helper.cpp
@@ -124,12 +124,12 @@ ExternalCommitHelper::ExternalCommitHelper(RealmCoordinator& parent, const Realm
     path = DB::get_core_file(config.path, DB::CoreFileType::Note);
     bool fifo_created = tessera::util::try_create_fifo(path, !temp_dir.empty() || !sys_temp_dir.empty());
     if (!fifo_created && !temp_dir.empty()) {
-        path = DB::get_core_file(util::format("%1realm_%2", temp_dir, std::hash<std::string>()(config.path)),
+        path = DB::get_core_file(util::format("%1tessera_%2", temp_dir, std::hash<std::string>()(config.path)),
                                  DB::CoreFileType::Note);
         fifo_created = tessera::util::try_create_fifo(path, !sys_temp_dir.empty());
     }
     if (!fifo_created && !sys_temp_dir.empty()) {
-        path = DB::get_core_file(util::format("%1realm_%2", sys_temp_dir, std::hash<std::string>()(config.path)),
+        path = DB::get_core_file(util::format("%1tessera_%2", sys_temp_dir, std::hash<std::string>()(config.path)),
                                  DB::CoreFileType::Note);
         tessera::util::create_fifo(path);
     }